+++ app/controller/debug/default.ts
@@ -5,7 +5,7 @@ export default class extends Controller {
   public async 'GET /test'() {
     const { ctx } = this;
     ctx.body = new Date();
-    ctx.model.sync({ force: true });
+
   }
 
 }