+++ app/service/user/project.ts
@@ -1,13 +1,12 @@
 import { Service } from 'egg';
-import { IFindOptions } from 'sequelize-typescript';
-import { v4 } from 'uuid';
+// import { v4 } from 'uuid';
+import { FindAndCountOptions } from 'sequelize';
 import { UserProject } from '../../model/user/project';
 
 export default class UserProjectService extends Service {
 
-
   async index(qry: { page: number, limit: number }) {
-    const opt:IFindOptions<UserProject> = {
+    const opt:FindAndCountOptions<UserProject> = {
       limit: qry.limit,
       offset: qry.limit * (qry.page - 1),
       attributes: [ 'uuid', 'name' ],
@@ -20,18 +19,21 @@ export default class UserProjectService extends Service {
   }
 
   async create(reqBody: { name: string }) {
-    const uv4 = v4();
-    console.log(uv4);
-    const uuidv4 = uv4.replace(/-/g, '');
-    const { uuid, name, owner_id } = await this.ctx.model.User.Project.create(
-      {
-        name: reqBody.name,
-        owner_id: this.ctx.passportUser.id,
-        uuid: uuidv4,
-      },
-    );
+    try {
+      const { uuid, name, owner_id } = await this.ctx.model.User.Project.create(
+        {
+          name: reqBody.name,
+          owner_id: this.ctx.passportUser.id,
+        },
+      );
 
-    return { uuid, name, owner_id };
+      return { status: 201, data: { uuid, name, owner_id } };
+    } catch (err) {
+      if (err.original?.errno === 1062) {
+        return { status: 409 }; // 409 "conflict"
+      }
+      throw err;
+    }
   }
 
 }