+++ test/app/.setup.ts
@@ -1,0 +1,8 @@
+import { app } from 'egg-mock/bootstrap';
+import factories, { reset } from './factories';
+
+
+before(() => factories(app));
+afterEach(async () => {
+  await reset();
+});