+++ test/app/controller/user/default.test.ts
@@ -1,21 +1,9 @@
 import { app } from 'egg-mock/bootstrap';
 import factory from 'factory-girl';
-import factories, { SeqName } from '../../factories';
+import { SeqName, ModelTypes, UserPassword } from '../../factories';
+import '../../.setup';
 
 describe('test/app/controller/user/default.test.ts', () => {
-
-
-  before(() => {
-    factories(app);
-  });
-
-  afterEach(async () => {
-    // clear database after each test case
-    factory.resetSequence(SeqName.UserAccount);
-    await Promise.all([
-      app.model.User.Account.destroy({ truncate: true, force: true }),
-    ]);
-  });
 
   describe('PUT /user/register', () => {
 
@@ -31,12 +19,10 @@ describe('test/app/controller/user/default.test.ts', () => {
     });
 
     it('409', async () => {
-      await factory.create(SeqName.UserAccount);
-      const account = 'account_1';
-      const password = 'password_1';
+      const { account } = await factory.create<ModelTypes.UserAccount>(SeqName.UserAccount);
       await app.httpRequest()
         .put(url)
-        .send({ account, password })
+        .send({ account, password: UserPassword })
         .expect(409);
     });
   });
@@ -56,24 +42,19 @@ describe('test/app/controller/user/default.test.ts', () => {
     });
 
     it("406 , invalidField is 'password'", async () => {
-      await factory.create(SeqName.UserAccount);
-
-      const account = 'account_1';
-      const password = '123456';
+      const { account, password } = await factory.create<ModelTypes.UserAccount>(SeqName.UserAccount);
       await app.httpRequest()
         .post(url)
-        .send({ account, password })
+        .send({ account, password: password + 'abc123' })
         .expect(406)
         .expect({ invalidField: 'password' });
     });
 
     it('302', async () => {
-      await factory.create(SeqName.UserAccount);
-      const account = 'account_1';
-      const password = 'password_1';
+      const { account } = await factory.create<ModelTypes.UserAccount>(SeqName.UserAccount);
       await app.httpRequest()
         .post(url)
-        .send({ account, password })
+        .send({ account, password: UserPassword })
         .expect(302);
     });